+++ src/webpack/production.js
@@ -1,4 +1,4 @@
-const merge = require('webpack-merge');
+const { merge } = require('webpack-merge');
 const TerserPlugin = require('terser-webpack-plugin');
 const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const shared = require('./shared');
@@ -45,5 +45,4 @@ module.exports = merge(shared(), {
       new OptimizeCSSAssetsPlugin({}),
     ],
   },
-  // devtool: 'source-map'
 });